+++ src/@components/OAuthPage/OAuthKakaoPage/index.tsx
@@ -1,4 +1,4 @@
-import Loading from "../@common/Loading";
+import Loading from "../../@common/Loading";
 
 export default function OAuthKakaoPage() {
   const code = new URL(window.location.href).searchParams.get("code");
