+++ src/@components/OAuthPage/OAuthNaverPage/index.tsx
@@ -1,4 +1,4 @@
-import Loading from "../@common/Loading";
+import Loading from "../../@common/Loading";
 
 export default function OAuthNaverPage() {
   const code = new URL(window.location.href).searchParams.get("code");
